Restructuring of transpose code
The implementation is cleanup up a bit and split into seperate a
subdirectory transpose and in seperate files.  The intention of these
changes is to have a seperate compilation step for the transpose related
code, resulting in libtranspose.so. Once this is done for other parts of
the code as well, the compilation of DedispPlan can be done without
nvcc.