Updated FFDPlan::execute
The main code is tested against the Python reference and produces the
same output. However, the code fails to correctly detect the pulsar in
testfdd. The input and output data seems ok when comparing the mean and
std in testfdd and execute. The execute method uses the dm list and spin
frequencies computed by Plan (this differs from the reference), but this
does not seem to fix it either.