Fix double-buffering of input for DedispPlan::execute_guru
Input buffers were overwritten before or during execution of the kernels
using these buffers. This did not show up when testing with testdedisp,
because it had only two gulps. The updated implementation does not need
the second thread (to copy the input), now that memcpy2D runs much
faster (in parallel).