Update batch_max in FDDGPUPlan::execute_gpu
The maximum batch size for ndm and nchan is now set dynamically based on
the parameters. The default values have been reduced. These changes should
improve the overall runtime in most cases, while this should also help
to prevent memory allocation errors that occur for a large number of DMs
and/or channels.